chore: remove unused code and consolidate duplicate implementations
- Remove unused imports from political_commentry_finder.py and api/server.py
- Remove unused IndividualReliabilityAnalysis class from utils/schemas/pydantic.py
- Remove unused TypedDicts (LegislationContent, IndividualStatementSummary, PoliticianStatementSummary) from utils/schemas/state.py
- Remove unused functions (search_property, get_entity_properties) from utils/tools/wikidata.py
- Remove unused get_agent_llm function and LLMModel enum from utils/llm
- Consolidate duplicate parse_mcp_result into shared utils/mcp/_shared.py
- Remove unused dependencies (tavily-python, python-multipart) from requirements.txt
