+++ utils/llm/__init__.py
@@ -1,13 +1,11 @@
-from .config import LLMModel, DEFAULT_TEMPERATURE, DEFAULT_MAX_TOKENS, DEFAULT_TIMEOUT
-from .factory import get_llm, get_mini_llm, get_agent_llm, get_structured_llm
+from .config import DEFAULT_TEMPERATURE, DEFAULT_MAX_TOKENS, DEFAULT_TIMEOUT
+from .factory import get_llm, get_mini_llm, get_structured_llm
 
 __all__ = [
-    "LLMModel",
     "DEFAULT_TEMPERATURE",
     "DEFAULT_MAX_TOKENS",
     "DEFAULT_TIMEOUT",
     "get_llm",
     "get_mini_llm",
-    "get_agent_llm",
     "get_structured_llm",
 ]
