+++ utils/schemas/state.py
@@ -33,34 +33,12 @@ class LegislationFinderState(BaseAgentState):
     reliable_legislation_sources: NotRequired[Annotated[list[str], operator.add]]
 
 
-class LegislationContent(TypedDict):
-    source: str
-    content: str
-    error: NotRequired[str]
-
-
-class IndividualStatementSummary(TypedDict):
-    """The information about a statement made by a politician for a specific legislative source."""
-
-    source: str
-    summary: str
-
-
-class PoliticianStatementSummary(TypedDict):
-    """The information about a politician that is dealing with legislation explored."""
-
-    name: str
-    statement_summaries: list[IndividualStatementSummary]
-
-
 class ChainData(TypedDict):
     """Data sent through the chain of AI components."""
 
     city: NotRequired[str]
     legislation_sources: NotRequired[str]
-    legislation_content: NotRequired[list[LegislationContent]]
     notes: NotRequired[str]
-    politician_public_statements: NotRequired[list[PoliticianStatementSummary]]
     legislation_summary: NotRequired[WriterOutput]
     markdown_report: NotRequired[str]
